Save shuffled object names by index.
This change enables for all platforms a Mac OS (prior to OS X) feature by which
shuffled object names are saved by index, and reshuffled on restore.  This
proved necessary to avoid mysterious segfaults on OS X and is probably a good
idea on all other platforms.

This change breaks save file compatibility.